Crate simulation

What is it?

A program to optimise the locations of police vehicles.

The program will take input files, which allow you to customise the area and police stations used for the experiment. It will then generate and improve solutions using a GA paired with a response simulation, which will provide an average response time to be used as a fitness function. An incident schedule is generated at the start of the program and every solution is tested on the same schedule to ensure fairness. This means one single run of this program will not provide a solution best to all circumstances.

Tuning GA

A variety of paremeters are available to tune the GA. Some relate to the GA process, i.e tournament size or mutation count. Others refer to the simulation, i.e timestep or the max number of cars available.

By default they are set to the parameters I have found most succesfull. However, this will change based on the area and constraints being used.

Set Up

Requires 3 files to be in place:

graph.json

police.json

probs.json

These files should be placed in a folder named x. Where x is the customisable variable ‘PLACE’ in read_data.rs.

To generate the files you must run the python script: map_data.py

Modules

data 🔒

Module containg the Data struct and it’s associated methods.

dijkstra_node 🔒

Module containing the DijkstraNode struct and it’s implementations to provide comparisons.

genetic 🔒

Module containing all the functions related to the genetic algorithm operation.

incident 🔒

Module containg the Incident struct and it’s associated methods.

node 🔒

Module containg the Node struct and it’s associated methods.

read_data 🔒

simulation 🔒

types 🔒

vehicle 🔒

Module containg the Vehicle struct and it’s associated methods.

Functions

main 🔒

Runs the entire GA process on input data pointed to by PLACE in read_data.rs.