Function simulation::genetic::evolve_pop

pub fn evolve_pop(
    solutions: &mut Vec<Vec<u8>>,
    num_to_select: u16,
    tournament_size: u16,
    eval_iterations: u8,
    spawn_stack: &Vec<Vec<Incident>>,
    graph: &HashMap<u64, Node>,
    base_locations: &Vec<u64>,
    route_cache: &mut HashMap<(u64, u64), f32>,
    timestep: f32,
    end_time: f32,
    max_cars: u16,
    num_of_mutations: u8,
    num_of_mutations_when_no_xover: u8,
    rngthread: &mut ThreadRng,
    unreachable_set: &mut HashSet<u64>,
    do_crossover: bool,
) -> f32

Evolves a population of solutions using:

Tournament selection

Crossover (and repair)

Mutation

Parameters

solutions - A Vec of solutions to evolve

num_to_select - The number of solutions to be selected from tournament selection.

tournament_size - The number of solutions to be entered into the tournament.

eval_iterations - The number of iterations to evaluate each solution for.

spawn_stack - The incident schedule.

graph - The graph of nodes.

base_locations - Vec of locations ids containg every police base.

route_cache - The cache of route calculations.

timestep - Timestep to run the simulation with.

end_time - Time to end simulation at.

max_cars - Maximum cars available.

num_of_mutations - The number of indexes to be mutated.

num_of_mutations_when_no_xover - The number of indexes to be mutated if do_crossover is False.

rngthread - A ThreadRng object to avoid constant reseeding.

unreachable_set - HashSet containg ids of every node that is unreachable.

do_crossover - Use crossover to evolve the population or not.