Function simulation::genetic::tournament_selection

pub fn tournament_selection(
    solutions: &Vec<Vec<u8>>,
    num_to_select: u16,
    tournament_size: u16,
    eval_iterations: u8,
    spawn_stack: &Vec<Vec<Incident>>,
    graph: &HashMap<u64, Node>,
    base_locations: &Vec<u64>,
    route_cache: &mut HashMap<(u64, u64), f32>,
    timestep: f32,
    end_time: f32,
    rngthread: &mut ThreadRng,
    unreachable_set: &mut HashSet<u64>,
    best_evaluated_fitness: &mut f32,
) -> Vec<usize>

Returns a list of indexes of the selected solutions after running a tournement selection with specified parameters.

Parameters

solutions - A Vec of solutions to run tournament selection on.

num_to_select - The number of solutions to be selected.

tournament_size - The number of solutions to be entered into the tournament.

eval_iterations - The number of iterations to evaluate each solution for.

spawn_stack - The incident schedule.

graph - The graph of nodes.

base_locations - Vec of locations ids containg every police base.

route_cache - The cache of route calculations.

timestep - Timestep to run the simulation with.

end_time - Time to end simulation at.

rngthread - A ThreadRng object to avoid constant reseeding.

unreachable_set - HashSet containg ids of every node that is unreachable.

best_evaluated_fitness - A reference to a Time value that will be updated with the best fitness discovered from the round of evaluations

Returns

A list of indexes of the solutions to be selected. Corresponds to the solutions parameter.