Struct simulation::incident::Incident

pub struct Incident {
    location: u64,
    service_time: f32,
    resolved_time: Option<f32>,
    creation_time: f32,
    valid: bool,
}

Represents an incident/crime that police vehicles can respond to in the simulation.

Fields
location: u64

The [Node] ID where the incident is taking place.

service_time: f32

How long will the police need to attend the incident for until it’s resolved.

resolved_time: Option<f32>

The time between the incident’s creation time and the police’s arrival. Will be None when not resolved and Some<Time> after police attendance.

creation_time: f32

The time the incident was created.

valid: bool

Is the incident reachable. Some incidents may spawn in disconnected areas of the graph with no police cars.

Implementations

impl Incident

pub fn new(location: u64, service_time: f32, creation_time: f32) -> Self

Returns an Incident object.

pub fn get_location(&self) -> u64

Returns the incident’s [Node] ID.

pub fn get_service_time(&self) -> f32

Returns the service time of the incident

pub fn get_resolved_time(&self) -> Option<f32>

Returns the time the incident was resolved in Some<Time> or None if the incident is yet to be resolved.

pub fn get_creation_time(&self) -> f32

Returns the creation time of the incident.

pub fn unreachable(&mut self)

Sets the incident as unreachable (invalid).

pub fn solved(&mut self, time: f32)

Sets the incident as resolved.

Parameters

time - The time the incident was resolved.

pub fn is_solved(&self) -> bool

Returns a boolean specifying if the incident is already solved or not.

pub fn is_valid(&self) -> bool

Returns a boolean specifying if the incident is reachable (valid) or not.