Struct simulation::signal::SignalConfig

pub struct SignalConfig {
    pub signal_noise: f32,
    pub gain_pattern: GainPattern,
    pub propagation_model: PropagationModel,
}

Configuration options for a signal model

Fields
signal_noise: f32

The amount of noise to add when generating new measurements with this model

gain_pattern: GainPattern

The pattern used for adjusting the signal gain

propagation_model: PropagationModel

The model used for calculating the base signal strength from distance

Methods

impl SignalConfig

fn new(signal_noise: f32, gain_pattern: GainPattern, propagation_model: PropagationModel) -> SignalConfig

Create a new signal model using the specified parameters