Module simulation::simulation

Module containing all methods related to the police response simulation.

The simulation can be explained as such:

Spawn new incidents using the spawn_stack incident schedule

Dispatch vehicles to incidents as efficiently as possible

Step the simulation and vehciles using the desired timstep

If the time < end_time: Go back to the start Else: Sum and return performance metrics

The simulation has been designed to run as efficiently as possible by using serveral concepts:

Hash Maps

Due to the large amount of items and info being stored, hash maps are used pretty much everywhere to achieve O(1) accessing.

Route Caching

Route calculations are done using Dijkstra and all knowledge gained from a Dijkstra run is cached in a HashMap. Before performing any calculations, this cache is checked to see if the calculation has been performed before and therefore doesn’t need to be done again.

Incident Plan

Predetermining the incident plan provides a fair test for all solutions. However it also has a couple performance boosts:

Makes route caching insanely powerful. If the incidents are always in the same place then journeys will be similar accross all solutions. Therefore caching is utilised to the maximum.

Removes need to randomly generate new incidents for every simulation.

Travel simplification

A vehicle will always store the cost of it’s latest journey. Then if a vehcile has just left base and now wants to return to base, it will just use this last travel cost value for the reverse journey rather than recalculating the new cost.

Functions

calc_travel_time 🔒

Calculates travel time from source to target or uses cache.

dijkstra 🔒

Performs a dijkstra algorithm. Terminates after target is located.

dispatching 🔒

Dispatches vehicles to active incidents.

evaluate

Evaluates a solution.

generate_incidents

Generates incidents based on probabilistic occurrence rates.

run 🔒

Runs the simulation once.

spawn_incidents 🔒

Adds incidents to the list of active ones. They are spawned from the spawn stack.

step_vehicles 🔒

Completes a timestep on each vehicle

update_route_cache 🔒

Adds new calculation knowledge into the route cache