Function simulation::simulation::evaluate

pub fn evaluate(
    solution: &Vec<u8>,
    iterations: u8,
    spawn_stack: &Vec<Vec<Incident>>,
    graph: &HashMap<u64, Node>,
    base_locations: &Vec<u64>,
    route_cache: &mut HashMap<(u64, u64), f32>,
    timestep: f32,
    end_time: f32,
    unreachable_set: &mut HashSet<u64>,
) -> f32

Evaluates a solution.

Parameters

solution - The solution to be evaluated.

iterations - The number of iterations for the solution to be evaluated. Redundant should always be one currently.

spawn_stack - Reference to a 2D Vec containing incidents to spawn at each timestep.

graph - The road network graph as a HashMap.

base_locations - A Vec of locations in which police vehicles can be based at.

route_cache - The current hash map of knowledge.

timestep - The simulation timestep.

end_time - The time to end simulation at.

unreachable_set - A HashSet of locations known to be unreachable that will be updated throughout the simulation.

Returns

types::Time - The average average response time.