Struct simulation::FilterFrame

pub struct FilterFrame {
    pub animal_state: AnimalState,
    pub particles: Vec<AnimalState>,
    pub particles_mean: AnimalState,
}

A frame representing a single target (i.e. animal) at a particular point in time

Fields
animal_state: AnimalState

The animal's true state

particles: Vec<AnimalState>

The current list of particles from the particle filter

particles_mean: AnimalState

An estimate of the animal's true state based on the mean value of all the particles