Struct simulation::SimulationConfig

pub struct SimulationConfig {
    pub uavs: Vec<UavConfig>,
    pub animals: Vec<AnimalConfig>,
    pub map_size: [f32; 2],
    pub num_particles: usize,
    pub particle_noise: f32,
}

Configuration options for a simulation

Fields
uavs: Vec<UavConfig>

The list of UAVs to simulate

animals: Vec<AnimalConfig>

The list of animals to simulate

map_size: [f32; 2]

The initial size of the map

num_particles: usize

The number of particles to use in the particle filter

particle_noise: f32

The amount of noise to use in the particle filter