Struct simulation::vehicle::Vehicle

pub struct Vehicle {
    base: u64,
    location: u64,
    secs_till_free: f32,
    name: String,
    last_travel: f32,
    just_come_from_base: bool,
}

Represents a police vehicle that can navigate the graph and be ordered to respond to [Incident].

Fields
base: u64

The [Node] ID of the vehicle’s base.

location: u64

The [Node] ID of the vehcile’s current location.

secs_till_free: f32

The amounts of seconds until the vehicle is free for new commands.

name: String

The name of the vehicle. Only ever used for logging or debugging.

last_travel: f32

The cost of the last travel the vehicle performed.

just_come_from_base: bool

A boolean that is ture if the last travel a vehicle made was directly from it’s base.

Implementations

impl Vehicle

pub fn new(base: u64, name: String) -> Self

Returns a new Vehicle object

Parameters

name - The name of the vehicle.

pub fn get_secs_till_free(&self) -> f32

Returns the number of seconds until the vehicle is free to recieve new commands.

pub fn get_location(&self) -> u64

Returns the node ID of the vehicle’s current location.

pub fn get_base(&self) -> u64

Returns the node ID of the vehicle’s base

pub fn get_name(&self) -> String

Returns the name of the vehcile

pub fn goto( &mut self, target: u64, travel_time: f32, incident_time: f32, ) -> Result<u64, &'static str>

Moves the vehicle to a new node.

Parameters

target - The node ID of the location to move to

travel_time The time it takes for the vehcile to travel to the new location.

incident_time If the vehicle is going to an incident how long will they have to attend? If it’s not an incident then set 0.

Returns

Result<types::Location, &'static str> - Either Ok or Err indicating if the move was successfull or not.

pub fn timestep(&mut self, timestep: f32)

Reduces the secs_till_free attribute by the indicated timestep.

Parameters

timestep - Time value to reduces secs_till_free by.

pub fn reset(&mut self)

Resets the vehcile back to it’s base. The last_travel value is used as a guess to how long the vehcile should be occupied for.

This method should only be used as a last resort if a vehicle becomes trapped.

pub fn last_move_from_base(&self) -> bool

Returns a boolean value indicating if the vehicle’s last move originated at it’s base.

pub fn get_last_travel(&self) -> f32

Returns the travel_cost of the vehicle’s last travel.