Crate simulation

This crate provides functionality for configuring and running a range-based tracking system.

Examples

Configuration options

Structs

AnimalConfig
Configuration options for an animal in a simulation

AnimalState
The state of an animal at a particular point in time

FilterFrame
A frame representing a single target (i.e. animal) at a particular point in time

Measurement
A measurement taken by a UAV at a particular point in time

Particle
The representation of a particle in the particle filter

SignalConfig
Configuration options for a signal model

SimulationConfig
Configuration options for a simulation

SimulationFrame
A snapshot of the simulation at a particular point in time

SimulationRunner
A structure for running a configured simulation

UavConfig
Configuration options for a UAV in the simulation

UavController
Configuration options for how the UAV is controlled

UavState
The state of the uav at a particular point in time